Various Changes. (1.5.0 Release)
Implemented full support for the recent territory changes:
    -    New values replace "control_xx" found in the worldStatus, FacilityControl and MetagameEvent events.
    -    Those wanting to still use the old method can still utilize the "total_xx" values, which use an identical method from pre-1.5.
Updated pom.xml version to 1.5.0 for release.
Refactored "Environment" into the static data package, and moved init inside of the static data manager.
Fixed an issue where the Query Manager singleton would not be initialized before the query request for static data.
Fixed a concurrency issue with the Query Manager.
Fixed a missing join in the main world query.
Fixed the query manager trying to use the reserved websocket environment for queries.
Fixed an issue with the static facility query, due to unreliable data from the facility_link collection from census.
Fixed an issue where the class/package name would be sent for the environment field, instead of the environment name itself.